chrome/content/mochitest/tests/test_requestList_memory.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Tests for request data</title>

  <link rel="stylesheet" type="text/css" href="/content/tests/SimpleTest/test.css" />

  <script type="text/javascript" src="/content/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/x-javascript;version=1.7" src="../httpd.js"></script>
  <script type="application/x-javascript;version=1.7" src="common.js"></script>
</head>
<body>
  <p id="display"></p>
  <div id="content" style="display: none">
    <iframe id="frame"></iframe>
  </div>
  <pre id="test">
  <script type="application/x-javascript;version=1.7">
    Cu.import(baseURL.spec + "RequestList.jsm");
    prepareFilterComponents();
    preparePrefs();

    let wrapNodeArray = Cc['@mozilla.org/array;1'].createInstance(Ci.nsIMutableArray);
    function wrapNode(node)
    {
      wrapNodeArray.appendElement(node, false);
      let result = wrapNodeArray.queryElementAt(0, Ci.nsISupports);
      wrapNodeArray.removeElementAt(0);
      return result;
    }

    let tests = [
      ["test1", "", sameAddressNodes1],
      ["test2", "", sameAddressNodes2],
      ["test3", "", sameAddressNodes3],
      ["test4", "", sameAddressNodes4],
      ["test1", "'+i+'", differentAddressNodes1],
      ["test2", "'+i+'", differentAddressNodes2],
      ["test3", "'+i+'", differentAddressNodes3],
      ["test4", "'+i+'", differentAddressNodes4],
    ];
    let currentTest = -1;
    let urlModifier = null;

    function start()
    {
      window.addEventListener("unload", stop, false);
      server.start(1234);

      server.registerPathHandler("/test1", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<script>for (var i = 0; i < 40; i++) new Image().src='test" + urlModifier + ".png';<" + "/script>";
        response.bodyOutputStream.write(body, body.length);
      });

      server.registerPathHandler("/test2", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<body onload='document.body.innerHTML=null;'><script>for (var i = 0; i < 40; i++) document.write('<iframe src=dummy" + urlModifier + "></iframe>')<" + "/script></body>";
        response.bodyOutputStream.write(body, body.length);
      });

      server.registerPathHandler("/test3", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<body><script>for (var i = 0; i < 40; i++) document.write('<iframe src=dummy" + urlModifier + "></iframe>')<" + "/script></body>";
        response.bodyOutputStream.write(body, body.length);
      });

      server.registerPathHandler("/test4", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<script>function addImage() { new Image().src='test" + urlModifier + ".png'; }for (var i = 0; i < 40; i++) addImage();<" + "/script>";
        response.bodyOutputStream.write(body, body.length);
      });

      runNextTest();
    }

    function runNextTest()
    {
      ++currentTest;
      if (currentTest >= tests.length)
      {
        $("frame").onload = null;
        SimpleTest.finish();
        return;
      }

      urlModifier = tests[currentTest][1];

      $("frame").onload = function() setTimeout(testPageLoaded, 0);
      $("frame").src = "http://127.0.0.1:1234/" + tests[currentTest][0];
    }

    function testPageLoaded()
    {
      tests[currentTest][2]();
      runNextTest();
    }

    function sameAddressNodes1()
    {
      let frameData = RequestList.getDataForWindow(wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 1, "Number of entries in the frame data");

      let entry = entries[0];
      is(entry.nodes.length, 40, "Number of associated nodes before garbage collection");

      top.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      is(entry.nodes.length, 0, "Number of associated nodes after garbage collection");
    }

    function sameAddressNodes2()
    {
      let frameData = RequestList.getDataForWindow(wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 1, "Number of entries in the frame data");

      let entry = entries[0];
      is(entry.nodes.length, 40, "Number of associated nodes before garbage collection");

      top.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      is(entry.nodes.length, 0, "Number of associated nodes after garbage collection");
    }

    function sameAddressNodes3()
    {
      let frameData = RequestList.getDataForWindow(wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 1, "Number of entries in the frame data");

      let entry = entries[0];
      is(entry.nodes.length, 40, "Number of associated nodes before garbage collection");

      top.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      is(entry.nodes.length, 40, "Number of associated nodes after garbage collection");
    }

    function sameAddressNodes4(event)
    {
      let frameData = RequestList.getDataForWindow(wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 1, "Number of entries in the frame data");

      let entry = entries[0];
      is(entry._nodes.length, 40, "Number of associated nodes before garbage collection");

      top.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();

      // Trigger list compression by adding a node while the counter has a high value
      entry._compactCounter = 100;
      window.frames[0].addImage();
      is(entry._nodes.length, 1, "Number of associated nodes after garbage collection");
    }

    function differentAddressNodes1()
    {
      let frameData = RequestList.getDataForWindow(wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 40, "Number of entries in the frame data before garbage collection");

      top.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      for each (let entry in entries)
        entry.lastUpdate = 0;

      entries = frameData.getAllLocations();
      is(entries.length, 0, "Number of entries in the frame data after garbage collection");
    }

    function differentAddressNodes2()
    {
      let frameData = RequestList.getDataForWindow(wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 40, "Number of entries in the frame data before garbage collection");

      top.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      for each (let entry in entries)
        entry.lastUpdate = 0;

      entries = frameData.getAllLocations();
      is(entries.length, 0, "Number of entries in the frame data after garbage collection");
    }

    function differentAddressNodes3()
    {
      let frameData = RequestList.getDataForWindow(wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 40, "Number of entries in the frame data before garbage collection");

      top.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      for each (let entry in entries)
        entry.lastUpdate = 0;

      entries = frameData.getAllLocations();
      is(entries.length, 40, "Number of entries in the frame data after garbage collection");
    }

    function differentAddressNodes4()
    {
      let frameData = RequestList.getDataForWindow(wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 40, "Number of entries in the frame data before garbage collection");

      top.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      for each (let entry in entries)
        entry.lastUpdate = 0;

      // Trigger list compression by adding a node while the counter has a high value
      frameData._compactCounter = 100;
      window.frames[0].addImage();

      entries = frameData.getAllLocations();
      is(entries.length, 1, "Number of entries in the frame data after garbage collection");
    }

    function stop()
    {
      server.stop();
    }

    let server = new nsHttpServer();
    SimpleTest.waitForExplicitFinish();
    addLoadEvent(start);
  </script>
  </pre>
</body>
</html>
